Actualizacion_Agosto

+++ database/migrations/2024_07_12_093730_create_consultas_table.php
@@ -8,21 +8,25 @@ class CreateConsultasTable extends Migration
 {
     public function up()
     {
-        Schema::create('consultas', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('cita_id');
-            $table->foreign('cita_id')->references('id')->on('citas')->onDelete('cascade');
-            $table->text('alergias')->nullable();
-            $table->text('receta')->nullable();
-            $table->foreignId('medicamento_id')->nullable()->constrained('medicamentos');
-            $table->integer('cantidad')->nullable();
-            $table->string('frecuencia')->nullable();
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('consultas')){
+            Schema::create('consultas', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('cita_id');
+                $table->foreign('cita_id')->references('id')->on('citas')->onDelete('cascade');
+                $table->text('alergias')->nullable();
+                $table->text('receta')->nullable();
+                $table->foreignId('medicamento_id')->nullable()->constrained('medicamentos');
+                $table->integer('cantidad')->nullable();
+                $table->string('frecuencia')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     public function down()
     {
+        Schema::disableForeignKeyConstraints();
         Schema::dropIfExists('consultas');
+        Schema::enableForeignKeyConstraints();
     }
 }
